pyproject.toml
[project]
name = "verisage"
version = "0.1.0"
description = "Verisage - Multi-LLM oracle for dispute resolution and fact verification"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "anthropic>=0.39.0",
    "google-genai>=0.3.0",
    "openai>=1.54.0",
    "httpx>=0.27.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "jinja2>=3.1.4",
    "python-dotenv>=1.0.1",
    "huey>=2.5.0",
    "x402>=0.2.1",
    "slowapi>=0.1.9",
    "oasis-rofl-client>=0.1.5",
    "coincurve>=20.0.0",
    "cdp-sdk>=0.21.0",
]

[dependency-groups]
dev = [
    "ruff>=0.8.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.pytest.ini_options]
pythonpath = "."
testpaths = ["tests"]
